requirements.txt
flask
requests
numpy==1.26.4
opencv-contrib-python==4.9.0.80
opencv-python
pyserial